VNX: persist queue cross device link error (#240)
The bug was reported: https://bugs.launchpad.net/cinder/+bug/1796827.

Storops supports the configured path and uses persist-queue to persist
the data to that path. However, the persist-queue is implemented to
write the data to a temp path, then leverages the atomic `os.rename` to
store the data to the user specified path.

The problem occurs when the temp path and the user specified path are on
two different devices. It raises `OSError: Invalid cross-device link`.

The fix is to make sure the temp path is no the same device of the user
specified one.